Fix asm defines generation
When integrating the compiler's stuff in the main Makefile, CLEANFILES
in Makefile.am was initialized two times. This resulted in asm_defines.h
not being cleaned with a `make clean` command.

Therefore, on my machine the file generated by the last GCC compilation
was kept and used.

On the other hand, this was rebuilt by clang on every testing
environment and other machines. This was breaking the final compilation,
because apparently CLANG looks at the text which is placed into an asm
volatile block before passing control to the assembler. We were abusing
the fact that GCC is not doing this, to generate incorrect assembly code
which was later manually parsed to generate a valid C file.

This has been fixed by replacing our invalid syntax with a line
beginning with a comment, so the assembler is happy and we can parse
anyhow.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>