Use a bitmap for lps_termination
Rather than using a vector of arrays, we rely on a bitmap to store the
termination decision by LPs.

Currently, this commit might introduce a slightly increased overhead in
`ccgs_compute_snapshot()` due to a slightly costly bitmap operation used
to set a bit, rather than copying a single boolean.

This is not a problem, as that part of core is about to go in a later
commit!

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>