Autonomic Power Management
This commit introduces the Power Management module in ROOT-Sim.
It is now possile to specify a power cap, and the run will be
monitored so as to enforce the maximum performance under this
power cap.

The code is ugly and the integration even worse. The technical
debt we have accumulated in this commit compares favorably with
the national debt.
But, at least, the initial tests are promising, and the integration
looks stable.

Co-authored-by: Stefano Conoci <conoci@dis.uniroma1.it>
Co-authored-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>
Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>