Remove unneeded fork()/wait()
In the compiler, a subcommand is spawned with system(). This system()
invocation was wrapped in a fork()/wait(), but system already forks()
before exec(), and is blocking with respect to the subcommand.

system() could be suboptimal for a number of reasons here, but
respecting its semantics and removing the unneeded fork() helps at
portability.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>