C11 threads custom implementation
MSVC does not implement C11 threads, because "they are not interested at
"implementing them".
We therefore included a slightly changed version of tinycthreads, to
account for MS idiosyncrasies with the standards. Indeed, C11 is 9 years
old, MS was in the standard committee, but they have no interest in
what they decided should go in the standard.
It's all very sad: we have to bloat the codebase just for a so-simple
thing.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>