Shared libraries, "visible" macro, topology cleanup
Enabled the build of ROOT-Sim shared libraries (still, our
custom compiler only statically links).

Symbols which are meant to be visible to the model developer
are now marked with a macro which changes their visibility.

Added a cleanup function for topologies; now the simulator is
valgrind friendly, once again!

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>