Re-added incremental state saving code, fixed no-op statistics
The code for the incremental state saving, on the side of the model
memory allocator, has been reintroduced.

Now, we do not leak memory and we do not open temporary files when
running without statistics enabled, also the simulation total time
is still correct.

Minor changes to the main CMakeLists.txt, to correctly build on
Ubuntu 20.04 with CMake 3.18.4.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>