Remove duplicated definitions and add a symbol with version
Some fundamental types are now defined in ROOT-Sim.h and it's
no longer necessary to define them in core.h.
Also, a variable keeping a textual representation of the version of
the core is introduced. It's not used now in the core (although it'll
be in some tests), but I consider this as an important bit to keep in
the binary. We'll see possible future usages.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>