Improved straggler detection
Added a new field in the processing context. It contains a lazily
updated bound value for the LP and it allows to detect straggles
without fetching past messages from memory.

A preliminary experimentation has shown that for large simulations
the reduction of time to completion can be as large as 10%.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>