Introduce configurable control messages
In a recent PR, we have phased out some libraries that were providing
to models functionalities that are not necessarily relevant to the core.
For proper development of these libraries, anyhow, they may need to
interact with the internals of the core, or extend part of their
functionalities.
Years ago we discussed the possibility to introduce a sort of SDK to
allow integrating additional layer over the core. This commit is a first
step in that direction.

Here, we are introducing the possibility to register handlers from
higher-level libraries, that are associated with proper control messages
that a library can exchange across nodes.
In this way, in a distributed simulation, a library can synchronize the
various nodes to configure the different instances.

I am not particularly happy with the reduced size of the payload, but
changing this would require to rework the antimessage extraction logic
which is something I don't want to do now. For the use case I am
running on, the current payload is sufficient.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>