Differentiate documentation target name across projects
We are concentrating the release actions in a single monorepo that
gathers together all the different projects.
For this to work, we need different unique target names across the
projects. This is already (mostly) true, except for the documentation
target. This commit prepends the name of the repo to the doc target,
for simplicity of management in the monorepo.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>