Remove AUTHORS file
We have the Contributors insights page on the project which can serve
well the purpose of the AUTHORS file, without having to maintain an
additional resource.
The drawback of this commit is that we lose memory of older
contributors, pre-GitHub era. At the same time, there is no longer any
code from them in the current code base.

Suggestions on this are welcome.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>