Dynamic LP assignment to threads/nodes
The code path involved in the routing of events across nodes/threads
now doesn't rely on a static assignment, but rather it will lookup
the correct node/thread dynamically when an event needs to be routed.
This is one of the first changes needed to implement LPs migration.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>