Renamed rid_t to tid_t, rid_t is now used for the LP context
The rid_t (resource id) type has been refactored to tid_t.
It was used only for thread anyways. The name resource id is now
used to identify the computational resource to which a LP is bound
that can either correspond to a thread or a remote MPI rank if the
LP is not locally hosted. This information is contained in the LP
struct, and should allow for non-blocking migrations and rebindings.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>