Preliminary integration of decision model
This is a preliminary integration of the various pieces of code.
The decision should be done based on the posted values.
I am resetting the time series data after a challenge outcome is
determined, although I am not 100% sure if it's in the right place.
The GPU is apparently not posting relevant measurements, have to
investigate if it's in the right place.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>