First revision of local LP rebinding
We introduce a periodic LP rebinding heuristic that uses measures of each LP
committed computational cost, and tries, using the Karmarkar-Karp differencing
algorithm, to assign LPs to threads to approximately minimize the workload
imbalance.

The logic is there, and it has been validated to do what is supposed to do;
still more reaserch is advisable to improve the heuristic that selects when
to rebind. In addition, the implementation could be improved, especially the
implementation of the Karmarkar-Karp algorithm.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>