Fix thread_affinity_set() on MacOS, improve error handling
As per Apple's infinite wisdom, the developer, clearly a
minus habens incapable of any understanding, is graciously
allowed to "hint" affinity preferences to the infallible
MacOS scheduler, which may even potentially consider them.

Refactor to provide more information to the user in case
thread affinity setting fails; distinguishing between the
case in which affinity is disabled or otherwise not
supported by the platform, and runtime errors.

Aligned windows thread affinity logic to linux.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>