Replace update-copyright.py with the update copyright action
The update copyright script didn't really belong to the code base
of the project. We had to manually update the copyright year,
and this typically happened very late in the year.

We now have a fancy action that's automatically triggered on Jan 1st
every year, to open a pull request with an automatic update of the
code base in `develop`.

Clearly, if we continue to spend years between a release and the other,
things may get misaligned easily.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>