Fix memory order in sync_thread_barrier()
In the previous version of the barrier the compiler could, in
theory, move some parts of the following critical sections
before the spin wait loop on the thread counter variable.

With the fix, this is not possible. We note though, that, on
Linux, both gcc 13.3 and clang 18 produce identical test
binaries.

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>