Reorganize the mm subsystem
This commit moves around functions and data structures, based on the
following core ideas:

- Buddy system stuff are confined in mm/buddy/ files
- model_allocator.c should implement everything that can be seen by
  the model or pertains to the LP state memory. It interacts with
  the buddy system, so the concept of "multi.c" is not necessary, nor
  pertains to mm/buddy/. Out of mm/ no one should explicitly see
  the buddy system.
- in mm/checkpoint we have everything that relates to checkpointing of
  the LP state. So we now have an `autonomic.c` unit, a `full.c` unit,
  and an `incremental.c` unit, that is currently only keeping old code.
  `mm/checkpoint/checkpoint.h` is the only header exposing all
  checkpointing-related stuff, so independently of what you need
  (autonomic/full/incremental), you can include this header.

I am convinced this refactor helps people go and find stuff based on
the organization of the tree. There could be a minimal penaly, as we
have lost one `static inline` for a function, but I am happy
tolerating a minimal performance penalty in favor of readability.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>