Overhaul of the termination module and many smaller changes
The termination module has been completely overhauled. Now
the approach is the correct one, in which we make use of
the GVT algorithm to ensure consistent and logically
synchronized termination of the simulation even in a
distributed simulation.

Also, I propose a new termination API (which will need
further iterations before stabilization) which, in my
mind, should substitute the handler-based termination
detection.

In addition:
- improved debug code related to event scheduling
- removed m_seq from struct lp_msg (merged with flags)
- improved safety of tagged pointers in process_ctx
- improved anti message matching in distributed runs
- some documentation efforts

Criticism and suggestions are always welcome!

Signed-off-by: Andrea Piccione <piccione@diag.uniroma1.it>